refactor: seasonalData 파일 import 경로 변경
