+++ src/Recruitment/Result/FinalResult.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
-import { generation, finalResult } from "../../seasonalData";
+import { generation, finalResult } from "../../data/seasonal/seasonalData";
 
 const FinalResult = ({ name, message, track }) => {
   return (
